requirements.txt
flask==2.3.2
ultralytics>=8.0.0        # YOLOv8 - training & inference
opencv-python-headless    # CV routines (use headless for servers)
torch>=2.0.0              # PyTorch - ultralytics will need this
numpy
pillow
werkzeug
gunicorn                 # for production
kaggle                   # to download datasets via CLI
